Enable to configure project specific settings for certain aspects
Currently many aspects are configured on a global level but actually it
would often be useful to configure them on a project level.
