+++ io.cucumber.eclipse.targetdefinition/cucumber.eclipse.targetdefinition.target
@@ -13,7 +13,7 @@
 			<dependency>
 				<groupId>io.cucumber</groupId>
 				<artifactId>gherkin</artifactId>
-				<version>32.1.1</version>
+				<version>33.0.0</version>
 				<type>jar</type>
 			</dependency>
 		</dependencies>
@@ -31,7 +31,7 @@ Export-Package:        *;version="${version}";-noimport:=true
 			<dependency>
 				<groupId>io.cucumber</groupId>
 				<artifactId>cucumber-java</artifactId>
-				<version>7.22.2</version>
+				<version>7.26.0</version>
 				<type>jar</type>
 			</dependency>
 		</dependencies>
@@ -49,7 +49,7 @@ Export-Package:        *;version="${version}";-noimport:=true
 			<dependency>
 				<groupId>io.cucumber</groupId>
 				<artifactId>cucumber-core</artifactId>
-				<version>7.22.2</version>
+				<version>7.26.0</version>
 				<type>jar</type>
 			</dependency>
 		</dependencies>
